all_fichiers_need/only_modif.py
# Génération des Parties Non Visibles avec IA

import torch
from torchvision.transforms import ToTensor, ToPILImage
from PIL import Image
import numpy as np

# Charger le modèle pré-entrainé (exemple VQ-VAE-2 ou GAN)
# Supposons que nous avons un modèle appelé "generative_model.pth"
model = torch.load("path/to/generative_model.pth")
model.eval()

def generate_hidden_parts(image_path):
    img = Image.open(image_path)
    img_tensor = ToTensor()(img).unsqueeze(0)  # Convertir en tenseur
    with torch.no_grad():
        generated_img_tensor = model(img_tensor)  # Générer l'image avec les parties cachées
    generated_img = ToPILImage()(generated_img_tensor.squeeze())
    return generated_img

if __name__ == "__main__":
    image_path = "path/to/your/image.png"
    generated_img = generate_hidden_parts(image_path)
    generated_img.save("generated_image.png")
