requirements.txt
flask
torch
torchvision
open3d
numpy
Pillow
opencv-python
timm